Reorganisation of the code
