+++ Excel_UI/Components/Adapter/CreateAdapter.cs
