+++ Excel_UI/Components/Adapter/Remove.cs
