+++ Excel_UI/Components/Engine/Explode.cs
