+++ Excel_UI/Components/Engine/GetInfo.cs
