+++ Excel_UI/Excel_UI.csproj
@@ -117,52 +117,55 @@
     <Compile Include="Addin\AddIn_OnOpen.cs" />
     <Compile Include="Addin\AddIn.cs" />
     <Compile Include="Addin\Ribbon.cs" />
+    <Compile Include="Caller\CallerFormula_Run.cs" />
+    <Compile Include="Caller\CallerFormula_SetItem.cs" />
+    <Compile Include="Caller\CallerFormula_Ribbon.cs" />
+    <Compile Include="Caller\CallerFormula_Register.cs" />
     <Compile Include="Helpers\ExcelSelectionHelper.cs" />
     <Compile Include="Helpers\ExcelCalculationManualHelper.cs" />
     <Compile Include="Helpers\ExcelEchoOffHelper.cs" />
     <Compile Include="Helpers\ArrayResizer.cs" />
-    <Compile Include="Methods\Properties.cs" />
-    <Compile Include="UI\Global\Project.cs" />
+    <Compile Include="Global\Project.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Properties\Resources.Designer.cs">
       <AutoGen>True</AutoGen>
       <DesignTime>True</DesignTime>
       <DependentUpon>Resources.resx</DependentUpon>
     </Compile>
-    <Compile Include="UI\Callers\Explode.cs" />
-    <Compile Include="UI\Components\Adapter\CreateAdapter.cs" />
-    <Compile Include="UI\Components\Adapter\CreateRequest.cs" />
-    <Compile Include="UI\Components\Adapter\Remove.cs" />
-    <Compile Include="UI\Components\Adapter\Execute.cs" />
-    <Compile Include="UI\Components\Adapter\Move.cs" />
-    <Compile Include="UI\Components\Adapter\Pull.cs" />
-    <Compile Include="UI\Components\Adapter\Push.cs" />
-    <Compile Include="UI\Components\Engine\Compute.cs" />
-    <Compile Include="UI\Components\Engine\Convert.cs" />
-    <Compile Include="UI\Components\Engine\Explode.cs" />
-    <Compile Include="UI\Components\Engine\FromJson.cs" />
-    <Compile Include="UI\Components\Engine\GetInfo.cs" />
-    <Compile Include="UI\Components\Engine\GetProperty.cs" />
-    <Compile Include="UI\Components\Engine\Modify.cs" />
-    <Compile Include="UI\Components\Engine\Query.cs" />
-    <Compile Include="UI\Components\Engine\SetProperty.cs" />
-    <Compile Include="UI\Components\Engine\ToJson.cs" />
-    <Compile Include="UI\Components\oM\CreateData.cs" />
-    <Compile Include="UI\Global\ComponentManager.cs" />
-    <Compile Include="UI\Global\ComponentUpgrader.cs" />
-    <Compile Include="UI\Templates\CallerValueListFormula.cs" />
-    <Compile Include="UI\Components\oM\CreateCustom.cs" />
-    <Compile Include="UI\Components\oM\CreateDictionary.cs" />
-    <Compile Include="UI\Components\oM\CreateEnum.cs" />
-    <Compile Include="UI\Components\oM\CreateObject.cs" />
-    <Compile Include="UI\Components\oM\CreateType.cs" />
-    <Compile Include="UI\Global\FormulaSearchMenu.cs" />
-    <Compile Include="UI\Templates\SelectorMenu_RibbonXml.cs" />
-    <Compile Include="UI\Templates\SingleOptionCallerFormula.cs" />
-    <Compile Include="UI\Templates\IExcelSelectorMenu.cs" />
-    <Compile Include="UI\Templates\SelectorMenuUtil.cs" />
-    <Compile Include="UI\Templates\CallerFormula.cs" />
-    <Compile Include="UI\Templates\FormulaDataAccessor.cs" />
+    <Compile Include="BHoM_UI_Overrides\Explode.cs" />
+    <Compile Include="Components\Adapter\CreateAdapter.cs" />
+    <Compile Include="Components\Adapter\CreateRequest.cs" />
+    <Compile Include="Components\Adapter\Remove.cs" />
+    <Compile Include="Components\Adapter\Execute.cs" />
+    <Compile Include="Components\Adapter\Move.cs" />
+    <Compile Include="Components\Adapter\Pull.cs" />
+    <Compile Include="Components\Adapter\Push.cs" />
+    <Compile Include="Components\Engine\Compute.cs" />
+    <Compile Include="Components\Engine\Convert.cs" />
+    <Compile Include="Components\Engine\Explode.cs" />
+    <Compile Include="Components\Engine\FromJson.cs" />
+    <Compile Include="Components\Engine\GetInfo.cs" />
+    <Compile Include="Components\Engine\GetProperty.cs" />
+    <Compile Include="Components\Engine\Modify.cs" />
+    <Compile Include="Components\Engine\Query.cs" />
+    <Compile Include="Components\Engine\SetProperty.cs" />
+    <Compile Include="Components\Engine\ToJson.cs" />
+    <Compile Include="Components\oM\CreateData.cs" />
+    <Compile Include="Global\ComponentManager.cs" />
+    <Compile Include="Global\ComponentUpgrader.cs" />
+    <Compile Include="Templates\CallerValueListFormula.cs" />
+    <Compile Include="Templates\FormulaDataAccessor.cs" />
+    <Compile Include="Templates\SingleOptionCallerFormula.cs" />
+    <Compile Include="Components\oM\CreateCustom.cs" />
+    <Compile Include="Components\oM\CreateDictionary.cs" />
+    <Compile Include="Components\oM\CreateEnum.cs" />
+    <Compile Include="Components\oM\CreateObject.cs" />
+    <Compile Include="Components\oM\CreateType.cs" />
+    <Compile Include="Global\FormulaSearchMenu.cs" />
+    <Compile Include="BHoM_UI_Overrides\SelectorMenu_RibbonXml.cs" />
+    <Compile Include="BHoM_UI_Overrides\IExcelSelectorMenu.cs" />
+    <Compile Include="BHoM_UI_Overrides\SelectorMenuUtil.cs" />
+    <Compile Include="Caller\CallerFormula.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="packages.config" />
