+++ Excel_UI/Global/ComponentManager.cs
